+++ app/Exceptions/Auth/AuthException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Exceptions\Auth;
+
+use RuntimeException;
+
+abstract class AuthException extends RuntimeException
+{
+}